Fix dnt script
